Making progress on model reorg
