+++ src/KMonad/Model/IO.hs
@@ -1,0 +1,10 @@
+module KMonad.Model.IO
+
+where
+
+import KMonad.Prelude
+import KMonad.Util.Logging
+import KMonad.Model.Types
+
+withModel :: LUIO m => ModelCfg -> Ctx r m ModelInterface
+withModel cfg = mkCtx $ \f -> undefined
