+++ src/KMonad/Pullchain.hs
@@ -4,4 +4,4 @@ where
 
 import KMonad.Pullchain.Action as X
 import KMonad.Pullchain.Button as X
-import KMonad.Pullchain.BEnv   as X
+import KMonad.Pullchain.Types  as X
